test: move more web contents specs (#20099)
