+++ shell/browser/api/event_emitter.h
@@ -107,15 +107,17 @@ class EventEmitter : public Wrappable<T> {
                      v8::Local<v8::Object> event,
                      Args&&... args) {
     DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
-    v8::Locker locker(isolate());
-    v8::HandleScope handle_scope(isolate());
-    EmitEvent(isolate(), GetWrapper(), name, event,
-              std::forward<Args>(args)...);
-    auto context = isolate()->GetCurrentContext();
+    // It's possible that |this| will be deleted by EmitEvent, so save anything
+    // we need from |this| before calling EmitEvent.
+    auto* isolate = this->isolate();
+    v8::Locker locker(isolate);
+    v8::HandleScope handle_scope(isolate);
+    auto context = isolate->GetCurrentContext();
+    EmitEvent(isolate, GetWrapper(), name, event, std::forward<Args>(args)...);
     v8::Local<v8::Value> defaultPrevented;
-    if (event->Get(context, StringToV8(isolate(), "defaultPrevented"))
+    if (event->Get(context, StringToV8(isolate, "defaultPrevented"))
             .ToLocal(&defaultPrevented)) {
-      return defaultPrevented->BooleanValue(isolate());
+      return defaultPrevented->BooleanValue(isolate);
     }
     return false;
   }
